Color helper functions exported to colorGenerator.js
